+++ sperl_var.h
@@ -6,7 +6,7 @@
 struct SPerl_var {
   SPerl_OP* op_name;
   SPerl_OP* op_my_var;
-  int32_t version;
+  int64_t version;
 };
 
 SPerl_VAR* SPerl_VAR_new(SPerl* sperl);
